+++ traffic_ops/testing/api/v4/deliveryservice_requests_test.go
@@ -408,10 +408,9 @@ func validatePutDSRequestFields(expectedResp map[string]interface{}) utils.CkReq
 
 func CreateTestDeliveryServiceRequests(t *testing.T) {
 	for _, dsr := range testData.DeliveryServiceRequests {
-		dsrV41 := dsr.Upgrade()
-		resetDS(dsrV41.Original)
-		resetDS(dsrV41.Requested)
-		respDSR, _, err := TOSession.CreateDeliveryServiceRequest(dsrV41, client.RequestOptions{})
+		resetDS(dsr.Original)
+		resetDS(dsr.Requested)
+		respDSR, _, err := TOSession.CreateDeliveryServiceRequest(dsr, client.RequestOptions{})
 		assert.NoError(t, err, "Could not create Delivery Service Requests: %v - alerts: %+v", err, respDSR.Alerts)
 	}
 }